Memoize &'static str tag inputs to 'static tag-derived outputs
I noticed that `xml_serde` regular expression matching took up a
significant fraction of my `cargo bench` profiling results and decided
to dig in.

First I refactored the regular expression into a `pub(crate) mod tag`,
with a `pub(crate) struct Tag<'a>` containing one `&'a str` field for
each regex capture. `Tag` can be `Copy`, and it can get conversion into
`xml::name::Name<'a>`. This was not a performance change, but it allowed
me to better see the dependency chains at work.

Deserializing is given field names as `&'static [&'static str]` which it
passes to the tag parser. Parsing a `&'static str` produces a
`Tag<'static>` -- and that value can be cached. No other string will
exist at the pointer address of that `&'static str` input string, which
makes it suitable for use as a map key. I therefore added
`Tag::from_static(&'static str) -> Tag<'static>` which adds a cache
around `Tag<'a>::new(&'a str) -> Tag<'a>`. I also added a cache in `de`
for the `&'static [&'static str]` itself: that entire `Field` slice
never modified after it's built, and it's already `'static` so the same
strategy applies. This led to a 50-60% reduction in runtime of my
element-heavy deserialization benchmarks.

Serializing is a bit more complex, since some of values which may need
matching against that regex are genuinely `String`s and not
`&'static str`s. I plumbed around `Cow<'static, str>` and added
`Tag::from_cow()`. If the `Cow::Borrowed` contains a static string, it
calls `Tag::from_static()` and takes advantage of the cache. If it does
not, it calls `Tag::new()` and matches the `String` as-is. This led to a
40-60% of my element-heavy serialization benchmarks.